+++ src/parser/Symbols.ts
@@ -113,6 +113,7 @@ export const CUT_SYMBOL = '✄';
 export const COPY_SYMBOL = '📚';
 export const PASTE_SYMBOL = '📋';
 export const CANCEL_SYMBOL = '⨉';
+export const CONFIRM_SYMBOL = '✓';
 export const ALL_SYMBOL = '⊞';
 export const ERASE_SYMBOL = '⌫';